Separate Unit test from TDD
