+++ src/main/java/org/secureauth/sarestapi/util/SAFactory.java
@@ -71,4 +71,5 @@ public class SAFactory {
 		}
 		return saAccess;
 	}
+
 }
