+++ src/test/resources/test.properties
@@ -1,4 +1,19 @@
+#saaccess
+idp.host=qaendpoint01.gosecureauth.com
+idp.port=443
+idp.realm=SecureAuth53
+idp.realm.api.application.id=000860c9324d4c79817d1edeeecce8ff
+idp.realm.api.application.key=e29e3f4f8a3150f579ea14bde65c9012aac65859ee1404d70c0ed655842b194b
+idp.ssl=true
+
 #User
-user.username=rtest1
+user.username=mhuber
 user.pin=1234
 user.password=1234
+
+#UserOATH-OTP
+user.oath.totp.factor.id=
+user.oath.totp.shared.key=
+user.oath.totp.length=6
+user.oath.totp.interval=
+user.otp.pin=000123
